Version 0.2.0: Add `define_gocd_tasks`
This version adds a DSL for defining common tasks: build, test, push,
and deploy. This lets us put common/current best practices in one place:

* Stop containers on clean
* Remove all containers on clobber
* Remove old build images but pull them before push/deploy
* Use GO_PIPELINE_NAME as project name when invoking `docker-compose`
* Use a volume for test artifacts

It should also dramatically simplify most of our build Rakefiles. Builds
that need to augment the default tasks can do so using normal Rake
semantics.

The :test task, in particular, doesn't do anything by default. Every
build needs to set up and run its tests explicitly in a :test task.
Also, if any artifacts need to be gathered from a test run, that should
be done within the :test task and not in a separate :test_results task.
The test volume helpers should make this easier.

The :deploy task doesn't invoke a lambda, it simply tags the image as
`:deployed` and pushes it. This gem provides `invoke_lambda` for builds
to use when augmenting the :deploy task.